Move tokens implementations to its own file
