+++ golem/transactions/ethereum/ethereumtransactionsystem.py
@@ -5,6 +5,7 @@ from ethereum.utils import privtoaddr
 
 from golem.ethereum import Client
 from golem.ethereum.paymentprocessor import PaymentProcessor
+from golem.ethereum.token import GNTToken
 from golem.report import report_calls, Component
 from golem.transactions.ethereum.ethereumpaymentskeeper \
     import EthereumAddress
@@ -38,9 +39,12 @@ class EthereumTransactionSystem(TransactionSystem):
 
         log.info("Node Ethereum address: " + self.get_payment_address())
 
+        client = Client(datadir, port, start_geth)
+        token = GNTToken(client)
         payment_processor = PaymentProcessor(
-            client=Client(datadir, port, start_geth),
+            client=client,
             privkey=node_priv_key,
+            token=token,
             faucet=True
         )
