+++ loggingconfig.py
@@ -78,6 +78,10 @@ LOGGING = {
             'level': 'INFO',
             'propagate': True,
         },
+        'golem.token': {
+            'level': 'INFO',
+            'propagate': True,
+        },
         'twisted': {
             'level': 'INFO',
             'propagate': True,
